sprites/requirements.txt
absl-py==0.1.10
asn1crypto==0.24.0
astor==0.7.1
atari-py==0.1.1
azure==1.0.3
azure-common==1.1.8
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==2.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==2.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
baselines==0.1.5
bleach==1.5.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
couchbase==2.3.4
cryptography==2.1.4
cycler==0.10.0
Cython==0.27.3
dask==1.0.0
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.7.1
docker==3.5.0
docker-pycreds==0.3.0
entrypoints==0.2.3
enum34==1.1.6
evdev==0.8.1
future==0.16.0
futures==3.1.1
gast==0.2.0
gitdb2==2.0.5
GitPython==2.1.11
glfw==1.5.1
grpcio==1.16.1
grpcio-tools==1.8.1
gym==0.10.5
h5py==2.8.0
html5lib==0.9999999
idna==2.6
imageio==2.2.0
imgaug==0.2.7
imutils==0.5.1
ipaddress==1.0.19
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.11.1
Jinja2==2.10
joblib==0.11
jsonpickle==1.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.0.8
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==2.2.3
mistune==0.8.4
mpi4py==3.0.0
mujoco-py==0.5.7
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.15.4
opencv-contrib-python==3.4.2.17
opencv-python==3.4.0.12
pandas==0.23.4
pandocfilters==1.4.2
parso==0.1.1
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
progressbar2==3.38.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.5.2
py4j==0.10.7
PyAudio==0.2.11
pycparser==2.18
pycrypto==2.6.1
pyglet==1.3.0
Pygments==2.2.0
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyprof2calltree==1.4.4
pyspark==2.4.0
python-dateutil==2.6.1
python-utils==2.3.0
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.18.4
-e git+https://github.com/openai/retro-contest.git@3cae0d0599f818fead1f16e008745f80cc0bbddb#egg=retro_contest_support&subdirectory=support
scandir==1.6
scikit-image==0.14.1
scipy==1.0.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.5
SQLAlchemy==1.3.12
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-gpu==1.8.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.19.5
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.52.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
yolk3k==0.9
zmq==0.0.0
